requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-debug-toolbar==4.4.2
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
kombu==5.4.0
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sentry-sdk==2.12.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
